requirements.gpu.txt
pillow==12.1.1
playwright==1.52.0
accelerate==1.13.0
huggingface_hub==1.8.0
tokenizers==0.22.2
safetensors==0.7.0
git+https://github.com/huggingface/transformers@8213e0d920d52cb00dcade16b6d1f6e952ac0a8c